fixed a few lil erroers

+++ Space Invaders/Space Invaders/LifeCount.cpp
@@ -21,7 +21,7 @@
 
 
 
-CLifeCount::CLifeCount(const float _iX, const float _iY)
+CLifeCount::CLifeCount(const int _iX, const int _iY)
 {
 	m_fX = _iX;
 	m_fY = _iY;
